refactor :: 가입 폼 관리 리팩토링
