+++ src/pages/management/index.ts
@@ -1,7 +1,7 @@
 export { default as InfoPage } from "./InfoPage";
 export { default as CampaignPage } from "./CampaignPage";
 export { default as ItemPage } from "./ItemPage/index";
-export { default as JoinFormPage } from "./JoinFormPage";
+export { default as JoinFormPage } from "./JoinFormPage/index";
 export { default as MachinePage } from "./MachinePage/index";
 export { default as NotiTalkPage } from "./NotiTalkPage";
 export { default as RentalPage } from "./RentalPage/index";
